# docs/openapi.yml
openapi: 3.0.3
info:
  title: ShelterFlex Backend API
  description: API for ShelterFlex backend service
  version: 0.1.0
  contact:
    name: ShelterFlex Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.shelterflex.io
    description: Production server
paths:
...
  /api/whistleblower/ratings:
    post:
      summary: Submit a whistleblower rating (tenant)
      description: Authenticated tenant submits a rating (1-5) and optional review text.
      operationId: createWhistleblowerRating
      tags:
        - Whistleblower
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWhistleblowerRatingRequest"
      responses:
        "201":
          description: Rating created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateWhistleblowerRatingResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/whistleblower/{id}/ratings:
    get:
      summary: List whistleblower ratings
      description: Public list of ratings for profile/dashboard display.
      operationId: listWhistleblowerRatings
      tags:
        - Whistleblower
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Ratings list
          content:
            application/json:
              schema:
                type: object
                required: [success, ratings]
                properties:
                  success:
                    type: boolean
                  ratings:
                    type: array
                    items:
                      $ref: "#/components/schemas/WhistleblowerRating"
        "400":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /api/whistleblower/{id}/ratings/aggregate:
    get:
      summary: Get whistleblower rating aggregate
      description: Public aggregate trust metrics (count, average, breakdown).
      operationId: getWhistleblowerRatingAggregate
      tags:
        - Whistleblower
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Aggregate metrics
          content:
            application/json:
              schema:
                type: object
                required: [success, aggregate]
                properties:
                  success:
                    type: boolean
                  aggregate:
                    $ref: "#/components/schemas/WhistleblowerRatingAggregate"
        "400":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
  /api/whistleblower/{id}/earnings:
    get:
      summary: Get whistleblower earnings
      description: Returns aggregated earnings totals and detailed history of individual rewards for a whistleblower
      operationId: getWhistleblowerEarnings
      tags:
        - Whistleblower
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the whistleblower
          schema:
            type: string
            minLength: 1
          example: "wb-123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: Earnings data retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsResponse"
              examples:
                mixedStatus:
                  summary: Whistleblower with mixed payout statuses
                  value:
                    totals:
                      totalNgn: 750000
                      pendingNgn: 450000
                      paidNgn: 300000
                      totalUsdc: 500
                      pendingUsdc: 300
                      paidUsdc: 200
                    history:
                      - rewardId: "rw-001"
                        listingId: "lst-456"
                        dealId: "deal-789"
                        amountNgn: 300000
                        amountUsdc: 200
                        status: "paid"
                        createdAt: "2024-01-15T10:30:00Z"
                        paidAt: "2024-01-20T14:45:00Z"
                      - rewardId: "rw-002"
                        listingId: "lst-457"
                        dealId: "deal-790"
                        amountNgn: 225000
                        amountUsdc: 150
                        status: "pending"
                        createdAt: "2024-01-10T08:15:00Z"
                      - rewardId: "rw-003"
                        listingId: "lst-458"
                        dealId: "deal-791"
                        amountNgn: 225000
                        amountUsdc: 150
                        status: "payable"
                        createdAt: "2024-01-05T16:20:00Z"
                allPending:
                  summary: Whistleblower with all pending rewards
                  value:
                    totals:
                      totalNgn: 600000
                      pendingNgn: 600000
                      paidNgn: 0
                      totalUsdc: 400
                      pendingUsdc: 400
                      paidUsdc: 0
                    history:
                      - rewardId: "rw-004"
                        listingId: "lst-459"
                        dealId: "deal-792"
                        amountNgn: 300000
                        amountUsdc: 200
                        status: "pending"
                        createdAt: "2024-02-01T12:00:00Z"
                      - rewardId: "rw-005"
                        listingId: "lst-460"
                        dealId: "deal-793"
                        amountNgn: 300000
                        amountUsdc: 200
                        status: "pending"
                        createdAt: "2024-01-28T09:30:00Z"
        "400":
          description: Invalid request - whistleblower ID format is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "Whistleblower ID is required"
                  details:
                    field: "id"
        "404":
          description: Whistleblower not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "Whistleblower not found"
                  details:
                    whistleblowerId: "wb-nonexistent"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Failed to retrieve earnings data"
  /api/whistleblower/listings:
    post:
      summary: Create a new listing
      description: |
        Create a new property listing. 
        
        Rules:
        - Address is required
        - Annual rent must be greater than 0
        - At least 3 photos are required
        - Maximum 2 listings per whistleblower per month
      operationId: createListing
      tags:
        - Whistleblower
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
      responses:
        '201':
          description: Listing created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - listing
                properties:
                  success:
                    type: boolean
                    example: true
                  listing:
                    $ref: '#/components/schemas/Listing'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingPhotos:
                  summary: Missing required photos
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid request data
                      details:
                        fields:
                          - field: photos
                            message: At least 3 photos are required
                invalidRent:
                  summary: Invalid annual rent
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid request data
                      details:
                        fields:
                          - field: annualRentNgn
                            message: Annual rent must be greater than 0
        '429':
          description: Monthly limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: CONFLICT
                  message: Monthly listing limit reached
                  details:
                    currentCount: 2
                    maxAllowed: 2
                    message: You have reached the maximum of 2 listings per month
        default:
          $ref: '#/components/responses/Error'

    get:
      summary: List listings
      description: Get a paginated list of listings with optional filters
      operationId: listListings
      tags:
        - Whistleblower
      parameters:
        - name: status
          in: query
          description: Filter by listing status
          required: false
          schema:
            $ref: '#/components/schemas/ListingStatus'
        - name: query
          in: query
          description: Search term for address, city, area, or description
          required: false
          schema:
            type: string
            example: Lagos
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: Listings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - listings
                  - pagination
                properties:
                  success:
                    type: boolean
                    example: true
                  listings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Listing'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /api/whistleblower/listings/{id}:
    get:
      summary: Get listing by ID
      description: Retrieve a single listing by its ID
      operationId: getListingById
      tags:
        - Whistleblower
      parameters:
        - name: id
          in: path
          description: Listing ID
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Listing retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - listing
                properties:
                  success:
                    type: boolean
                    example: true
                  listing:
                    $ref: '#/components/schemas/Listing'
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: NOT_FOUND
                  message: Listing with ID '550e8400-e29b-41d4-a716-446655440000' not found
        default:
          $ref: '#/components/responses/Error'
...
  /api/admin/whistleblower/listings:
    get:
      summary: List whistleblower listings (admin)
      description: |
        Returns a paginated list of whistleblower listings for admin review.
        Defaults to `status=pending_review` when no status filter is provided.
      operationId: adminListWhistleblowerListings
      tags:
        - Admin
      parameters:
        - name: status
          in: query
          description: Filter by listing status (defaults to pending_review)
          required: false
          schema:
            type: string
            enum: [pending_review, approved, rejected, rented]
            default: pending_review
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of listings
          content:
            application/json:
              schema:
                type: object
                required:
                  - listings
                  - pagination
                properties:
                  listings:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminListing'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/Error'
...
components:
  schemas:
...
    CreateWhistleblowerRatingRequest:
      type: object
      required: [whistleblowerId, dealId, rating]
      properties:
        whistleblowerId:
          type: string
          minLength: 1
        dealId:
          type: string
          minLength: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
          nullable: true
          maxLength: 2000
    WhistleblowerRating:
      type: object
      required:
        - ratingId
        - whistleblowerId
        - tenantId
        - dealId
        - rating
        - createdAt
      properties:
        ratingId:
          type: string
          format: uuid
        whistleblowerId:
          type: string
        tenantId:
          type: string
        dealId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateWhistleblowerRatingResponse:
      type: object
      required: [success, rating]
      properties:
        success:
          type: boolean
        rating:
          $ref: "#/components/schemas/WhistleblowerRating"
    WhistleblowerRatingAggregate:
      type: object
      required: [whistleblowerId, count, average, breakdown]
      properties:
        whistleblowerId:
          type: string
        count:
          type: integer
        average:
          type: number
        breakdown:
          type: object
          additionalProperties:
            type: integer
...
    AdminListing:
      type: object
      required:
        - listingId
        - whistleblowerId
        - address
        - bedrooms
        - bathrooms
        - annualRentNgn
        - photos
        - status
        - createdAt
        - updatedAt
      properties:
        listingId:
          type: string
          format: uuid
        whistleblowerId:
          type: string
        address:
          type: string
        city:
          type: string
        area:
          type: string
        bedrooms:
          type: integer
          minimum: 0
        bathrooms:
          type: integer
          minimum: 0
        annualRentNgn:
          type: integer
          minimum: 1
        description:
          type: string
        photos:
          type: array
          items:
            type: string
            format: uri
        status:
          type: string
          enum: [pending_review, approved, rejected, rented]
        reviewedBy:
          type: string
          description: ID of the admin who reviewed this listing
        reviewedAt:
          type: string
          format: date-time
          description: Timestamp of the review decision
        rejectionReason:
          type: string
          description: Reason provided when the listing was rejected
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ListingStatus:
      type: string
      enum:
        - pending_review
        - approved
        - rejected
        - rented
      example: pending_review
    CreateListingRequest:
      type: object
      required:
        - whistleblowerId
        - address
        - bedrooms
        - bathrooms
        - annualRentNgn
        - photos
      properties:
        whistleblowerId:
          type: string
          description: ID of the whistleblower creating the listing
          example: wb-001
        address:
          type: string
          description: Full address of the property
          minLength: 1
          example: 123 Main Street, Ikeja, Lagos
        city:
          type: string
          description: City where the property is located
          example: Lagos
        area:
          type: string
          description: Specific area or neighborhood
          example: Ikeja
        bedrooms:
          type: integer
          description: Number of bedrooms
          minimum: 0
          example: 2
        bathrooms:
          type: integer
          description: Number of bathrooms
          minimum: 0
          example: 2
        annualRentNgn:
          type: integer
          description: Annual rent in Nigerian Naira
          minimum: 1
          example: 1200000
        description:
          type: string
          description: Optional description of the property
          example: Beautiful 2-bedroom apartment with modern amenities
        photos:
          type: array
          description: Array of photo URLs (minimum 3, maximum 20)
          minItems: 3
          maxItems: 20
          items:
            type: string
            format: uri
            example: https://example.com/photo1.jpg
    Listing:
      type: object
      required:
        - listingId
        - whistleblowerId
        - address
        - bedrooms
        - bathrooms
        - annualRentNgn
        - photos
        - status
        - createdAt
        - updatedAt
      properties:
        listingId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        whistleblowerId:
          type: string
          example: wb-001
        address:
          type: string
          example: 123 Main Street, Ikeja, Lagos
        city:
          type: string
          nullable: true
          example: Lagos
        area:
          type: string
          nullable: true
          example: Ikeja
        bedrooms:
          type: integer
          example: 2
        bathrooms:
          type: integer
          example: 2
        annualRentNgn:
          type: integer
          example: 1200000
        description:
          type: string
          nullable: true
          example: Beautiful 2-bedroom apartment
        photos:
          type: array
          items:
            type: string
            format: uri
          example:
            - https://example.com/photo1.jpg
            - https://example.com/photo2.jpg
            - https://example.com/photo3.jpg
        status:
          $ref: '#/components/schemas/ListingStatus'
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:00:00Z
        rejectionReason:
          type: string
          nullable: true
          description: Reason for rejection (only present if status is rejected)
          example: Insufficient photo quality
        dealId:
          type: string
          format: uuid
          nullable: true
          description: ID of the deal this listing is linked to (only present if status is rented)
          example: 550e8400-e29b-41d4-a716-446655440001
...
    Pagination:
      type: object
      required:
        - total
        - page
        - pageSize
        - totalPages
      properties:
        total:
          type: integer
          description: Total number of items
          example: 42
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of items per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
...
    EarningsResponse:
      type: object
      required:
        - totals
        - history
      properties:
        totals:
          $ref: "#/components/schemas/EarningsTotals"
        history:
          type: array
          items:
            $ref: "#/components/schemas/EarningsHistoryItem"
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - TOO_MANY_REQUESTS
                - NOT_FOUND
                - CONFLICT
                - LISTING_ALREADY_RENTED
                - SOROBAN_ERROR
                - INTERNAL_ERROR
              example: NOT_FOUND
            message:
              type: string
              example: "Resource not found"
            details:
              type: object
              description: Optional structured context for the error
              additionalProperties: true
...
  responses:
    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "TOO_MANY_REQUESTS"
              message: "Rate limit exceeded"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "Request validation failed"
              details:
                fieldErrors:
                  - field: "email"
                    message: "Invalid email address"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
...